+++ clang-tools-extra/test/clang-tidy/checkers/modernize/avoid-c-style-cast.cpp
@@ -368,3 +368,12 @@ void functional_casts() {
 
   throw S2(5.0f);
 }
+void f(int *);
+
+void f(double *);
+void test_nullptr_cast() {
+  f((int*)nullptr);
+}
+// CHECK-MESSAGES: :[[@LINE-1]]:5: warning: C-style casts are discouraged [modernize-avoid-c-style-cast]
+// CHECK-FIXES: f(static_cast<int*>(nullptr));
+