+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -634,6 +634,10 @@ Changes in existing checks
   on Windows when the check was enabled with a 32-bit :program:`clang-tidy`
   binary.
 
+- Improved :doc:`modernize-avoid-c-style-cast  
+  <clang-tidy/checks/modernize/avoid-c-style-cast>` by providing correct fixes 
+  for C-style casts of nullptr.
+  
 - Improved :doc:`modernize-use-override
   <clang-tidy/checks/modernize/use-override>` by fixing an issue where
   the check would sometimes suggest inserting ``override`` in an invalid